requirements.txt
discord.py>=2.0.0
python-dotenv
deepl
aiofiles
aiohttp
openai
aiosqlite
pytest-asyncio
cachetools
requests
deep-translator
numpy>=2.0.0,<2.3.0
psutil>=5.9.0

# OCR for ranking screenshot processing
Pillow>=9.0.0
pytesseract>=0.3.10
easyocr>=1.7.1
craft-text-detector>=0.4.3

